apps/rating-worker/requirements.txt
boto3==1.42.94
psycopg2-binary==2.9.12
watchtower==3.4.0
